+++ encodings/dict/src/builders/mod.rs
@@ -1,16 +1,11 @@
 use bytes::BytesDictBuilder;
 use primitive::PrimitiveDictBuilder;
-use vortex_array::arrays::{
-    ConstantArray, PrimitiveArray, PrimitiveEncoding, VarBinArray, VarBinViewArray,
-};
-use vortex_array::compute::try_cast;
-use vortex_array::stats::Stat;
+use vortex_array::arrays::{PrimitiveArray, VarBinArray, VarBinViewArray};
+use vortex_array::compress::downscale_integer_array;
 use vortex_array::variants::PrimitiveArrayTrait;
-use vortex_array::vtable::EncodingVTable;
-use vortex_array::{Array, ArrayExt, ArrayRef, ToCanonical};
-use vortex_dtype::{DType, PType, match_each_native_ptype};
-use vortex_error::{VortexExpect, VortexResult, vortex_bail};
-use vortex_scalar::Scalar;
+use vortex_array::{Array, ArrayExt, ArrayRef};
+use vortex_dtype::match_each_native_ptype;
+use vortex_error::{VortexResult, vortex_bail};
 
 use crate::DictArray;
 
@@ -50,100 +45,4 @@ pub fn dict_encode(array: &dyn Array) -> VortexResult<DictArray> {
         );
     }
     Ok(dict_array)
-}
-
-/// Downscale a primitive array to the narrowest PType that fits all the values.
-pub fn downscale_integer_array(array: ArrayRef) -> VortexResult<ArrayRef> {
-    if !array.is_encoding(PrimitiveEncoding.id()) {
-        // This can happen if e.g. the array is ConstantArray.
-        return Ok(array);
-    }
-    if array.is_empty() {
-        return Ok(array);
-    }
-    let array = array
-        .as_opt::<PrimitiveArray>()
-        .vortex_expect("Checked earlier");
-
-    let min = array.statistics().compute_stat(Stat::Min)?;
-    let max = array.statistics().compute_stat(Stat::Max)?;
-
-    let (Some(min), Some(max)) = (min, max) else {
-        // This array but be all nulls.
-        return Ok(
-            ConstantArray::new(Scalar::null(array.dtype().clone()), array.len()).into_array(),
-        );
-    };
-
-    // If we can't cast to i64, then leave the array as its original type.
-    // It's too big to downcast anyway.
-    let Ok(min) = i64::try_from(&min) else {
-        return Ok(array.to_array());
-    };
-    let Ok(max) = i64::try_from(&max) else {
-        return Ok(array.to_array());
-    };
-
-    downscale_primitive_integer_array(array.clone(), min, max).map(|a| a.into_array())
-}
-
-/// Downscale a primitive array to the narrowest PType that fits all the values.
-fn downscale_primitive_integer_array(
-    array: PrimitiveArray,
-    min: i64,
-    max: i64,
-) -> VortexResult<PrimitiveArray> {
-    if min < 0 || max < 0 {
-        // Signed
-        if min >= i8::MIN as i64 && max <= i8::MAX as i64 {
-            return try_cast(
-                &array,
-                &DType::Primitive(PType::I8, array.dtype().nullability()),
-            )?
-            .to_primitive();
-        }
-
-        if min >= i16::MIN as i64 && max <= i16::MAX as i64 {
-            return try_cast(
-                &array,
-                &DType::Primitive(PType::I16, array.dtype().nullability()),
-            )?
-            .to_primitive();
-        }
-
-        if min >= i32::MIN as i64 && max <= i32::MAX as i64 {
-            return try_cast(
-                &array,
-                &DType::Primitive(PType::I32, array.dtype().nullability()),
-            )?
-            .to_primitive();
-        }
-    } else {
-        // Unsigned
-        if max <= u8::MAX as i64 {
-            return try_cast(
-                &array,
-                &DType::Primitive(PType::U8, array.dtype().nullability()),
-            )?
-            .to_primitive();
-        }
-
-        if max <= u16::MAX as i64 {
-            return try_cast(
-                &array,
-                &DType::Primitive(PType::U16, array.dtype().nullability()),
-            )?
-            .to_primitive();
-        }
-
-        if max <= u32::MAX as i64 {
-            return try_cast(
-                &array,
-                &DType::Primitive(PType::U32, array.dtype().nullability()),
-            )?
-            .to_primitive();
-        }
-    }
-
-    Ok(array)
 }
