+++ vortex-array/src/compress.rs
@@ -1,88 +1,105 @@
-// FIXME(ngates): move this file into the compressor
-use vortex_error::VortexResult;
+use vortex_dtype::{DType, PType};
+use vortex_error::{VortexExpect, VortexResult};
+use vortex_scalar::Scalar;
 
-use crate::aliases::hash_set::HashSet;
-use crate::stats::PRUNING_STATS;
-use crate::{Array, ArrayRef, EncodingId};
+use crate::arrays::{ConstantArray, PrimitiveArray, PrimitiveEncoding};
+use crate::compute::try_cast;
+use crate::stats::Stat;
+use crate::vtable::EncodingVTable;
+use crate::{Array, ArrayExt, ArrayRef, ToCanonical};
 
-/// Extendable compression interface, allowing implementations to explore different choices.
-pub trait CompressionStrategy {
-    /// Compress input array.
-    fn compress(&self, array: &dyn Array) -> VortexResult<ArrayRef>;
-
-    /// A set of the IDs of the encodings the compressor can choose from.
-    fn used_encodings(&self) -> HashSet<EncodingId>;
-}
-
-/// Check that compression did not alter the length of the validity array.
-pub fn check_validity_unchanged(arr: &dyn Array, compressed: &dyn Array) {
-    let _ = arr;
-    let _ = compressed;
-    #[cfg(debug_assertions)]
-    {
-        use vortex_error::VortexExpect;
-
-        let old_validity = arr
-            .validity_mask()
-            .vortex_expect("failed to compute validity")
-            .len();
-        let new_validity = compressed
-            .validity_mask()
-            .vortex_expect("failed to compute validity ")
-            .len();
-
-        debug_assert!(
-            old_validity == new_validity,
-            "validity length changed after compression: {old_validity} -> {new_validity}\n From tree {} To tree {}\n",
-            arr.tree_display(),
-            compressed.tree_display()
-        );
+/// Downscale a primitive array to the narrowest PType that fits all the values.
+pub fn downscale_integer_array(array: ArrayRef) -> VortexResult<ArrayRef> {
+    if !array.is_encoding(PrimitiveEncoding.id()) {
+        // This can happen if e.g. the array is ConstantArray.
+        return Ok(array);
     }
-}
-
-/// Check that compression did not alter the dtype
-pub fn check_dtype_unchanged(arr: &dyn Array, compressed: &dyn Array) {
-    let _ = arr;
-    let _ = compressed;
-    #[cfg(debug_assertions)]
-    {
-        debug_assert!(
-            arr.dtype() == compressed.dtype(),
-            "Compression changed dtype: {} -> {}\nFrom array: {}Into array {}",
-            arr.dtype(),
-            compressed.dtype(),
-            arr.tree_display(),
-            compressed.tree_display(),
-        );
+    if array.is_empty() {
+        return Ok(array);
     }
+    let array = array
+        .as_opt::<PrimitiveArray>()
+        .vortex_expect("Checked earlier");
+
+    let min = array.statistics().compute_stat(Stat::Min)?;
+    let max = array.statistics().compute_stat(Stat::Max)?;
+
+    let (Some(min), Some(max)) = (min, max) else {
+        // This array but be all nulls.
+        return Ok(
+            ConstantArray::new(Scalar::null(array.dtype().clone()), array.len()).into_array(),
+        );
+    };
+
+    // If we can't cast to i64, then leave the array as its original type.
+    // It's too big to downcast anyway.
+    let Ok(min) = i64::try_from(&min) else {
+        return Ok(array.to_array());
+    };
+    let Ok(max) = i64::try_from(&max) else {
+        return Ok(array.to_array());
+    };
+
+    downscale_primitive_integer_array(array.clone(), min, max).map(|a| a.into_array())
 }
 
-// Check that compression preserved the statistics.
-pub fn check_statistics_unchanged(arr: &dyn Array, compressed: &dyn Array) {
-    let _ = arr;
-    let _ = compressed;
-    #[cfg(debug_assertions)]
-    {
-        use crate::stats::StatsProviderExt;
+/// Downscale a primitive array to the narrowest PType that fits all the values.
+fn downscale_primitive_integer_array(
+    array: PrimitiveArray,
+    min: i64,
+    max: i64,
+) -> VortexResult<PrimitiveArray> {
+    if min < 0 || max < 0 {
+        // Signed
+        if min >= i8::MIN as i64 && max <= i8::MAX as i64 {
+            return try_cast(
+                &array,
+                &DType::Primitive(PType::I8, array.dtype().nullability()),
+            )?
+            .to_primitive();
+        }
 
-        // Run count merge_ordered assumes that the run is "broken" on each chunk, which is a useful estimate but not guaranteed to be correct.
-        for (stat, value) in arr.statistics().to_owned().into_iter() {
-            if let Some(dtype) = stat.dtype(compressed.dtype()) {
-                let compressed_scalar = compressed.statistics().get_scalar(stat, &dtype);
-                debug_assert_eq!(
-                    compressed_scalar,
-                    Some(value.clone().into_scalar(dtype)),
-                    "Compression changed {stat} from {value} to {:?}",
-                    compressed_scalar.as_ref(),
-                );
-            }
+        if min >= i16::MIN as i64 && max <= i16::MAX as i64 {
+            return try_cast(
+                &array,
+                &DType::Primitive(PType::I16, array.dtype().nullability()),
+            )?
+            .to_primitive();
+        }
+
+        if min >= i32::MIN as i64 && max <= i32::MAX as i64 {
+            return try_cast(
+                &array,
+                &DType::Primitive(PType::I32, array.dtype().nullability()),
+            )?
+            .to_primitive();
+        }
+    } else {
+        // Unsigned
+        if max <= u8::MAX as i64 {
+            return try_cast(
+                &array,
+                &DType::Primitive(PType::U8, array.dtype().nullability()),
+            )?
+            .to_primitive();
+        }
+
+        if max <= u16::MAX as i64 {
+            return try_cast(
+                &array,
+                &DType::Primitive(PType::U16, array.dtype().nullability()),
+            )?
+            .to_primitive();
+        }
+
+        if max <= u32::MAX as i64 {
+            return try_cast(
+                &array,
+                &DType::Primitive(PType::U32, array.dtype().nullability()),
+            )?
+            .to_primitive();
         }
     }
-}
 
-/// Eagerly compute certain statistics (i.e., pruning stats plus UncompressedSizeInBytes) for an array.
-/// This function is intended to be called in compressors, immediately before compression occurs.
-pub fn compute_precompression_stats(arr: &dyn Array) -> VortexResult<()> {
-    arr.statistics().compute_uncompressed_size_in_bytes();
-    arr.statistics().compute_all(PRUNING_STATS).map(|_| ())
+    Ok(array)
 }
