+++ vortex-btrblocks/src/patches.rs
@@ -1,8 +1,8 @@
 use vortex_array::Array;
 use vortex_array::arrays::ConstantArray;
+use vortex_array::compress::downscale_integer_array;
 use vortex_array::compute::scalar_at;
 use vortex_array::patches::Patches;
-use vortex_dict::builders::downscale_integer_array;
 use vortex_error::VortexResult;
 
 /// Compresses the given patches by downscaling integers and checking for constant values.
