+++ vortex-btrblocks/src/string.rs
@@ -2,7 +2,7 @@ use vortex_array::aliases::hash_set::HashSet;
 use vortex_array::arrays::VarBinViewArray;
 use vortex_array::{Array, ArrayRef, ToCanonical};
 use vortex_dict::DictArray;
-use vortex_dict::builders::{dict_encode, downscale_integer_array};
+use vortex_dict::builders::dict_encode;
 use vortex_error::{VortexExpect, VortexResult};
 use vortex_fsst::{fsst_compress, fsst_train_compressor};
 
@@ -194,9 +194,8 @@ impl Scheme for DictScheme {
         }
 
         // Find best compressor for codes and values separately
-        let downscaled_codes = downscale_integer_array(dict.codes().to_array())?.to_primitive()?;
         let compressed_codes = IntCompressor::compress(
-            &downscaled_codes,
+            &dict.codes().to_primitive()?,
             is_sample,
             allowed_cascading - 1,
             &[crate::integer::DictScheme.code()],
