+++ vortex-btrblocks/src/temporal.rs
@@ -1,9 +1,9 @@
 //! Specialized compressor for DateTimeParts metadata.
 
 use vortex_array::arrays::TemporalArray;
+use vortex_array::compress::downscale_integer_array;
 use vortex_array::{Array, ArrayRef, ToCanonical};
 use vortex_datetime_parts::{DateTimePartsArray, TemporalParts, split_temporal};
-use vortex_dict::builders::downscale_integer_array;
 use vortex_error::VortexResult;
 
 use crate::integer::IntCompressor;
